+++ tests/MudBlazor.Markdown.Tests/Components/MarkdownComponentTests/MarkdownComponentLinksShould.cs
@@ -1,7 +1,7 @@
 ﻿using Markdig.Syntax.Inlines;
 using MyNihongo.Option;
 
-namespace MudBlazor.Markdown.Tests.MarkdownComponentTests;
+namespace MudBlazor.Markdown.Tests.Components.MarkdownComponentTests;
 
 public sealed class MarkdownComponentLinksShould : MarkdownComponentTestsBase
 {
@@ -35,10 +35,12 @@ public sealed class MarkdownComponentLinksShould : MarkdownComponentTestsBase
 	public void NavigateWhenLinkClicked()
 	{
 		const string value =
-@"[東京](#tokyo)
-[札幌](#sapporo)
-# Tokyo
-## Sapporo";
+			"""
+			[東京](#tokyo)
+			[札幌](#sapporo)
+			# Tokyo
+			## Sapporo
+			""";
 
 		using var fixture = CreateFixture(value);
 
@@ -64,31 +66,36 @@ public sealed class MarkdownComponentLinksShould : MarkdownComponentTestsBase
 	public void OverrideAllLinks()
 	{
 		const string value =
-@"[absolute](https://www.google.co.jp/)
-[relative](/tokyo)
-[id](#edogawa)";
+			"""
+			[absolute](https://www.google.co.jp/)
+			[relative](/tokyo)
+			[id](#edogawa)
+			""";
 
 		const string expected =
-@"<article class='mud-markdown-body'>
-	<p class='mud-typography mud-typography-body1'>
-		<a rel='noopener noreferrer' href='overriddenhttps://www.google.co.jp/' target='_blank' class='mud-typography mud-link mud-primary-text mud-link-underline-hover mud-typography-body1'>
-			absolute
-		</a>
-		<br />
-		<a href='overridden/tokyo' class='mud-typography mud-link mud-primary-text mud-link-underline-hover mud-typography-body1'>
-			relative
-		</a>
-		<br />
-		<a href='overridden#edogawa' class='mud-typography mud-link mud-primary-text mud-link-underline-hover mud-typography-body1'>
-			id
-		</a>
-	</p>
-</article>";
+			"""
+			<article class='mud-markdown-body'>
+				<p class='mud-typography mud-typography-body1'>
+					<a rel='noopener noreferrer' href='overriddenhttps://www.google.co.jp/' target='_blank' class='mud-typography mud-link mud-primary-text mud-link-underline-hover mud-typography-body1'>
+						absolute
+					</a>
+					<br />
+					<a href='overridden/tokyo' class='mud-typography mud-link mud-primary-text mud-link-underline-hover mud-typography-body1'>
+						relative
+					</a>
+					<br />
+					<a href='overridden#edogawa' class='mud-typography mud-link mud-primary-text mud-link-underline-hover mud-typography-body1'>
+						id
+					</a>
+				</p>
+			</article>
+			""";
 
 		var @override = Optional<Func<LinkInline, string?>?>.Of(Override);
 
 		using var fixture = CreateFixture(value, overrideLinkUrl: @override);
 		fixture.MarkupMatches(expected);
+		return;
 
 		static string Override(LinkInline x) =>
 			"overridden" + x.Url;
@@ -99,16 +106,19 @@ public sealed class MarkdownComponentLinksShould : MarkdownComponentTestsBase
 	{
 		const string value = @"![img](/tokyo/sky-tree.png)";
 		const string expected =
-@"<article class='mud-markdown-body'>
-	<p class='mud-typography mud-typography-body1'>
-		<img src='overridden/tokyo/sky-tree.png' alt='img' class='mud-image object-fill object-center mud-elevation-25 rounded-lg'>
-	</p>
-</article>";
+			"""
+			<article class='mud-markdown-body'>
+				<p class='mud-typography mud-typography-body1'>
+					<img src='overridden/tokyo/sky-tree.png' alt='img' class='mud-image object-fill object-center mud-elevation-25 rounded-lg'>
+				</p>
+			</article>
+			""";
 
 		var @override = Optional<Func<LinkInline, string?>?>.Of(Override);
 
 		using var fixture = CreateFixture(value, overrideLinkUrl: @override);
 		fixture.MarkupMatches(expected);
+		return;
 
 		static string Override(LinkInline x) =>
 			"overridden" + x.Url;
