+++ tests/MudBlazor.Markdown.Tests/Components/MarkdownComponentTests/MarkdownComponentMathShould.cs
@@ -1,4 +1,4 @@
-﻿namespace MudBlazor.Markdown.Tests.MarkdownComponentTests;
+﻿namespace MudBlazor.Markdown.Tests.Components.MarkdownComponentTests;
 
 public sealed class MarkdownComponentMathShould : MarkdownComponentTestsBase
 {
@@ -8,9 +8,11 @@ public sealed class MarkdownComponentMathShould : MarkdownComponentTestsBase
 		const string value = "$\\sqrt {2}$";
 
 		const string expected =
-@"<article class='mud-markdown-body'>
-	<p class='mud-typography mud-typography-body1'>\(\sqrt {2}\)</p>
-</article>";
+			"""
+			<article class='mud-markdown-body'>
+				<p class='mud-typography mud-typography-body1'>\(\sqrt {2}\)</p>
+			</article>
+			""";
 
 		using var fixture = CreateFixture(value);
 		fixture.MarkupMatches(expected);
@@ -22,9 +24,11 @@ public sealed class MarkdownComponentMathShould : MarkdownComponentTestsBase
 		const string value = "$$\\sqrt {2}$$";
 
 		const string expected =
-@"<article class='mud-markdown-body'>
-	<p class='mud-typography mud-typography-body1'>$$\sqrt {2}$$</p>
-</article>";
+			"""
+			<article class='mud-markdown-body'>
+				<p class='mud-typography mud-typography-body1'>$$\sqrt {2}$$</p>
+			</article>
+			""";
 
 		using var fixture = CreateFixture(value);
 		fixture.MarkupMatches(expected);
