+++ tests/MudBlazor.Markdown.Tests/MudBlazor.Markdown.Tests.csproj
@@ -37,4 +37,9 @@
 		<Compile Include=".extensions\*.cs" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Components\" />
+	  <Folder Include="Utils\" />
+	</ItemGroup>
+
 </Project>
