+++ tests/MudBlazor.Markdown.Tests/Utils/Extensions/CodeBlockExTests/CodeBlockExTestsBase.cs
@@ -1,6 +1,6 @@
 using System.Text;
 
-namespace MudBlazor.Markdown.Tests.Extensions.CodeBlockExTests;
+namespace MudBlazor.Markdown.Tests.Utils.Extensions.CodeBlockExTests;
 
 public abstract class CodeBlockExTestsBase
 {
